Fix search for denom function
The internal binary search fn expects f(i+1) to be true if f(i) is true, so it essentially finds the first index of the matching string, if it exists, or the first index of the string that is higher lexicographically. We must thus use compare >= 0 in the binary search fn and then check if the returned value is not our denom.